moved to use now tickerplot package - basic testing
Fairly big logical change to move to use tickerplot as a separate
package that provides a lot of utilites. So that package can be used
wherever applicable.

Also DB name is now provided run-time and hence not hard-coded

Renamed get_stocks_nse2.py -> get_stocks_nse.py
